refactor: create pages folder and add pokemon search page into it
